Move const to an isolated file
When const is in background.ts, errors occur by importing background.ts
